Use signet client properly
